GROVE-149: --template and --templateRelease
Instead of --templateName / -n, which was something like "Vue" or "React", we
now support --template / -t, which is something like
"grove-vue-template" or "grove-react-template".

We also change --templateVersion / -v to --templateRelease / -r. This is
currently resolved to a git branch or tag, but eventually could mean
something else. This change leaves open `-v` to mean `--verbose`.