GROVE-413: add --keepGit as option
By default, remove git and initialize a new git repo with an "Initial
commit by grove cli".

If --keepGit is specified, rename `origin` remote to `upstream`.

Also reduced the "config" that gets passed to createNew(). This should
only be things actually going into the configuration of the project.
Also passing through the `program` options, so things like `keepGit` are
accessible separately from the `config`.

Also shortened the output on `grove new` a bit.